frontend/requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
extra-streamlit-components==0.1.71
fastapi==0.112.2
folium==0.17.0
gitdb==4.0.11
GitPython==3.1.43
hasher==2.0.1
idna==3.8
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
repl==1.0
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.4
streamlit==1.38.0
streamlit-authenticator==0.3.3
streamlit_folium==0.22.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
xyzservices==2024.6.0
tinydb-4.8.0
streamlit-js-1.0.8